+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Bugfixes
 
  - Fix issue with conditionals not triggering when measurement values are 0 ([#387](https://github.com/kizniche/mycodo/issues/387))
+ - Fix issue with settings Output PWM duty cycles
  - Fix issues with Atlas UART module ([#382](https://github.com/kizniche/mycodo/issues/382))
 
 
@@ -45,7 +46,6 @@
 ### Bugfixes
 
  - Fix inability to change Input Period ([#393](https://github.com/kizniche/mycodo/issues/393))
- - Fix issue with settings Output PWM duty cycles
  - Fix Exception while reading the GPIO pin of Edge Conditional ([#387](https://github.com/kizniche/mycodo/issues/387))
 
 ### Miscellaneous